Added Swipe Gesture to the Card
Animated a Single Card to move according to user's touch.
Added rotation to the card according to user's change in x direction i.e. interpolated rotation to linear drag in x direction.
Limited the rotation to -120deg to 120deg considering the user's mobile screen size by making use of Dimensions object of react-native.
Added Spring animation to reset the position of Card on touch release.
Forced swipe the card out of the screen if user dragged the card more than quater of the screen in the respective directions.